# server/docs/openapi.yaml
openapi: '3.0.3'
info:
  title: EasyConvert API
  description: LLM-based resume parsing API with multi-tenant support
  version: '1.0.0'
  contact:
    name: EasyConvert Team
servers:
  - url: http://localhost:3001
    description: Development server
paths:
...
  /api/v1/parse:
    post:
      summary: Synchronous single file parse
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF or DOCX resume file (max 10MB)
      responses:
        '200':
          description: Parse result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/parse/async:
    post:
      summary: Asynchronous parse, returns jobId
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: pending
                  message:
                    type: string
  /api/v1/parse/{jobId}:
    get:
      summary: Query async job status
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
  /api/v1/parse/batch:
    post:
      summary: Batch parse (max 20 files)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 20
      responses:
        '202':
          description: Batch queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobIds:
                    type: array
                    items:
                      type: string
                      format: uuid
                  totalFiles:
                    type: integer
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: 'API Key: Authorization: Bearer ec_...'
  schemas:
    Resume:
      type: object
      required: [basics, work, education]
      properties:
        basics:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
            title:
              type: string
            location:
              type: string
        work:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              position:
                type: string
              startDate:
                type: string
              endDate:
                type: string
              responsibilities:
                type: array
                items:
                  type: string
              achievements:
                type: array
                items:
                  type: string
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
              degree:
                type: string
              major:
                type: string
              startDate:
                type: string
              endDate:
                type: string
        skills:
          type: array
          items:
            type: string
        summary:
          type: string
    ParseResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            resume:
              $ref: '#/components/schemas/Resume'
            classification:
              type: object
            difficulty:
              type: string
            fromCache:
              type: boolean
            hash:
              type: string
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        fileName:
          type: string
        result:
          $ref: '#/components/schemas/Resume'
        error:
          type: string
        processingTime:
          type: integer
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              retryAfter:
                type: integer